Spell the CreateBVConst value parameter uint64_t
The three CreateBVConst overloads that take a value took it as an
"unsigned long long int". That is 64 bits everywhere, so unlike the
other conversions in this series it was never a portability bug -- it
is a spelling change, for consistency with the rest of the interface.

It is not a no-op, though. On LP64 uint64_t is unsigned long, a type
distinct from unsigned long long despite the identical width, so the
mangled names change:

  _ZN3stp6STPMgr13CreateBVConstEjy  ->  _ZN3stp6STPMgr13CreateBVConstEjm

and likewise for Cpp_interface and NodeFactory. That is an ABI break on
Linux and macOS. On i386 and Windows, where long is 32 bits, uint64_t
already is unsigned long long and the symbols are unchanged.

Taken deliberately: the next release carries other ABI breaks.

The headers now include <cstdint> rather than relying on picking it up
transitively, and test_fprewrites drops a cast that is now redundant.

+++ include/stp/cpp_interface.h
@@ -29,6 +29,7 @@ THE SOFTWARE.
 #include "stp/NodeFactory/NodeFactory.h"
 #include "stp/Util/Attributes.h"
 #include "extlib-unordered-dense/ankerl/unordered_dense.h"
+#include <cstdint>
 #include <map>
 #include <memory>
 #include <string>
@@ -236,7 +237,7 @@ public:
                                    int bit_width);
   DLL_PUBLIC ASTNode CreateBVConst(const char* const strval, int base);
   DLL_PUBLIC ASTNode CreateBVConst(unsigned int width,
-                                   unsigned long long int bvconst);
+                                   uint64_t bvconst);
   DLL_PUBLIC ASTNode CreateRMConst(unsigned mode);
   DLL_PUBLIC ASTNode CreateSourceSymbol(const char* name,
                                         const SourceSort& source_sort);
